+++ R/DSC_BICO.R
@@ -86,7 +86,3 @@ BICO_R$methods(
   }
 )
 
-DSC_registry$set_entry(name = "DSC_BICO",
-  DSC_Micro = TRUE, DSC_Macro = FALSE,
-  description = "BICO - Fast computation of k-means coresets")
-