+++ R/DSC_DBSCAN.R
@@ -111,7 +111,3 @@ DBSCAN$methods(
   }
 )
 
-DSC_registry$set_entry(name = "DSC_DBSCAN",
-  DSC_Micro = FALSE, DSC_Macro = TRUE,
-  description = "DBSCAN Reclustering")
-