+++ R/DSC_DBSTREAM.R
@@ -443,7 +443,3 @@ get_cluster_assignments <- function(x) {
   structure(last, ids = ids)
 }
 
-
-DSC_registry$set_entry(name = "DSC_DBSTREAM",
-  DSC_Micro = TRUE, DSC_Macro = FALSE,
-  description = "DBSTREAM - density-based stream clustering with shared-density-based reclustering.")