+++ R/DSC_evoStream.R
@@ -156,7 +156,3 @@ evoStream_R$methods(
   }
 )
 
-DSC_registry$set_entry(name = "DSC_evoStream",
-  DSC_Micro = TRUE, DSC_Macro = FALSE,
-  description = "evoStream - Evolutionary Stream Clustering")
-